wip / do not review / CI testing - generic hasattr on base variable
[ghstack-poisoned]
